data-ingestion/requirements.txt
# Web scraping
scrapy>=2.10.0
scrapy-playwright>=0.0.25
playwright>=1.32.0

# Enhanced scraping tools for anti-detection
selenium>=4.15.0
undetected-chromedriver>=3.5.0
fake-useragent>=1.4.0
scrapy-rotating-proxies>=0.6.2
scrapy-user-agents>=0.1.1

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Storage and databases
pymongo>=4.3.0
redis>=4.5.0

# Kafka for real-time data streaming
kafka-python>=2.0.2

# Big data tools
# hdfs3>=0.3.1  # Commented out due to Python 3.12 compatibility issues
# Alternative: Use pydoop or hdfs for HDFS operations
kafka-python>=2.0.2

# Utility libraries
requests>=2.28.0
python-dotenv>=1.0.0
schedule>=1.2.0

# Monitoring and logging
prometheus-client>=0.16.0
psutil>=5.9.0